requirements.txt
fastapi
uvicorn
streamlit
chromadb
sentence-transformers
google-generativeai
openai
groq
requests
python-dotenv
